refactor: refactor ui components
